requirements.txt
psutil==5.9.4
opencv-python==4.7.0.72
numpy==1.23.5
matplotlib==3.7.1
tensorboard==2.12.0
torchsummary==1.5.1
addict==2.4.0
yapf==0.32.0
imageio==2.26.1
colorlog==6.7.0
scipy==1.10.1
timm==0.8.17.dev0
h5py
pytorch-lightning
imageio
scikit-image
einops
pillow
knockknock
clearml
thop
fvcore